+++ src/main/java/org/example/model/Order.java
@@ -9,7 +9,7 @@ public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "order_id")
-    private Long orderId;
+    private int orderId;
 
     @ManyToOne
     @JoinColumn(name = "customer_id")
@@ -27,11 +27,11 @@ public class Order {
 
     //Getters and setters
 
-    public Long getOrderId() {
+    public int getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Long orderId) {
+    public void setOrderId(int orderId) {
         this.orderId = orderId;
     }
